voice-assistant/requirements.txt
# Voice Assistant Python Dependencies
# Install with: pip install -r requirements.txt

# Core AI/ML
faster-whisper==1.0.3
wyoming==1.5.3
ollama==0.3.3

# Web/Network
aiohttp==3.10.5
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0

# Data Processing
numpy==1.26.4
pydantic==2.9.2

# Audio Processing
soundfile==0.12.1
webrtcvad==2.0.10

# Configuration & Utilities
python-dotenv==1.0.1
colorlog==6.8.2
pyyaml==6.0.1
psutil==6.1.0
